+++ inStrain/controller.py
@@ -5,6 +5,7 @@ from ._version import __version__
 import os
 import sys
 import h5py
+import psutil
 import logging
 import argparse
 import pandas as pd
@@ -115,9 +116,9 @@ class ProfileController():
 ***************************************************
         """
         logging.info(message)
-        logging.debug('Checkpoint {0} start'.format('filter_reads'))
+        logging.debug('Checkpoint {1} start RAM is {0}'.format(psutil.virtual_memory()[1], 'filter_reads'))
 
-        global s2l # make ths global so we can access it later.
+        # global s2l # make ths global so we can access it later.
 
         # Parse the args
         bam = args.bam
@@ -164,14 +165,14 @@ class ProfileController():
         vargs['s2s'] = s2s
         vargs['s2p'] = s2p
 
-        logging.debug('Checkpoint {0} end'.format('filter_reads'))
+        logging.debug('Checkpoint {1} end RAM is {0}'.format(psutil.virtual_memory()[1], 'filter_reads'))
         message = """\
 ***************************************************
 .:: inStrain profile Step 2. Profile scaffolds ::..
 ***************************************************
         """
         logging.info(message)
-        logging.debug('Checkpoint {0} start'.format('profile_scaffolds'))
+        logging.debug('Checkpoint {1} start RAM is {0}'.format(psutil.virtual_memory()[1], 'profile_scaffolds'))
 
         Sprofile = inStrain.profileUtilities.profile_bam(bam, FAdb, Rdic, **vargs)
 
@@ -194,7 +195,7 @@ class ProfileController():
         # Run the rest of things
         args.IS = Sprofile.location
 
-        logging.debug('Checkpoint {0} end'.format('profile_scaffolds'))
+        logging.debug('Checkpoint {1} end RAM is {0}'.format(psutil.virtual_memory()[1], 'profile_scaffolds'))
         # See if you can profile genes as well
         message = """\
 ***************************************************
@@ -203,10 +204,10 @@ class ProfileController():
         """
         logging.info(message)
         if args.gene_file != None:
-            logging.debug('Checkpoint {0} start'.format('profile_genes'))
+            logging.debug('Checkpoint {1} start RAM is {0}'.format(psutil.virtual_memory()[1], 'profile_genes'))
             args.IS = Sprofile.location
             Controller().profile_genes_operation(args)
-            logging.debug('Checkpoint {0} end'.format('profile_genes'))
+            logging.debug('Checkpoint {1} end RAM is {0}'.format(psutil.virtual_memory()[1], 'profile_genes'))
         else:
             logging.info('Nevermind! You didnt include a genes file')
 
@@ -218,10 +219,10 @@ class ProfileController():
         """
         logging.info(message)
         if not args.skip_genome_wide:
-            logging.debug('Checkpoint {0} start'.format('genome_wide'))
+            logging.debug('Checkpoint {1} start RAM is {0}'.format(psutil.virtual_memory()[1], 'genome_wide'))
             args.IS = Sprofile.location
             Controller().genome_wide_operation(args)
-            logging.debug('Checkpoint {0} end'.format('genome_wide'))
+            logging.debug('Checkpoint {1} end RAM is {0}'.format(psutil.virtual_memory()[1], 'genome_wide'))
         else:
             logging.info('Nevermind! You chose to skip genome_wide')
 
@@ -233,11 +234,11 @@ class ProfileController():
         """
         logging.info(message)
         if not args.skip_plot_generation:
-            logging.debug('Checkpoint {0} start'.format('making_plots'))
+            logging.debug('Checkpoint {1} start RAM is {0}'.format(psutil.virtual_memory()[1], 'making_plots'))
             args.IS = Sprofile.location
             args.plots = 'a'
             Controller().plot_operation(args)
-            logging.debug('Checkpoint {0} end'.format('making_plots'))
+            logging.debug('Checkpoint {1} end RAM is {0}'.format(psutil.virtual_memory()[1], 'making_plots'))
         else:
             logging.info('Nevermind! You chose to skip making plots')